requirements.txt
appdirs==1.4.2
appnope==0.1.0
certifi==2017.1.23
cffi==1.9.1
cryptography==1.7.2
decorator==4.0.11
Fabric3==1.13.1.post1
flake8==3.3.0
future==0.16.0
idna==2.4
ipython==5.3.0
ipython-genutils==0.1.0
mccabe==0.6.1
packaging==16.8
paramiko==2.1.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.3
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.5.0
Pygments==2.2.0
pyparsing==2.1.10
python-telegram-bot==5.3.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.20
wcwidth==0.1.7
